update: create gates and auth restrictions to events and attendees

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -17,6 +17,7 @@ class AttendeeController extends Controller
 
     public function __construct(array $relations = ['user']) {
         $this->relations = $relations;
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
     }
 
     /**
@@ -56,6 +57,7 @@ class AttendeeController extends Controller
      */
     public function destroy(string $event, Attendee $attendee)
     {
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
